fixup! resize cpufreq
After commit e631c2c used, they don't adjust properly [1]

Cpufreq is a measure to optimize power during communication between "ROM and kernel". The kernel already uses it and we should adapt it from the device tree. Because it is the default pulse level on bengal (The 2.2Ghz level can be used by the kernel without adjusting to 2.2Ghz level)